requirements.txt
gradio==4.26.0
fastapi==0.110.1
pydantic==2.7.0
huggingface_hub==0.23.4
tensorflow-cpu==2.15.0
scikit-learn==1.6.1
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
sqlalchemy==2.0.30
pymysql==1.1.1
bcrypt==4.1.3
python-dotenv==1.0.1
reportlab==4.2.2
joblib==1.4.2
scipy==1.13.1
Pillow==10.3.0